Enhanced identification of HOST_IP and eth device - V2
Update of the prior commit to simplify the identification (only run
`ifconfig' once and save the results in a shell variable for easy
access).

Also, add the ability to search the `nova.conf' file for the host IP
address and interface name if this is the second time the script has
been run (nova networking removes the IP address from the public
interface, leaving the `nova.conf' file as the only place to find it
again).